changed poap providers name
