+++ group-generators/generators/sismo-diggers/index.ts
@@ -13,7 +13,7 @@ const generator: GroupGenerator = {
     // This group is constituted by all the users who have a sismo poap
     // of the following event:
 
-    const poapProvider = new dataProviders.AggregatedPoapSubgraphProvider();
+    const poapProvider = new dataProviders.PoapSubgraphProvider();
 
     const zikiPoapOwners = await poapProvider.queryEventsTokenOwners({
       eventIds: [
